разбираюсь со статистикой
