+++ docker-compose.yml
@@ -1,7 +1,7 @@
 version: '3.1'
 services:
   stats-server:
-    build: ./stats-parent/stats-service
+    build: stats/stats-service
     image: stats-service
     container_name: stats-service
     ports:
@@ -10,10 +10,11 @@ services:
       - stats-db
     environment:
       - SPRING_DATASOURCE_URL=jdbc:postgresql://stats-db:5432/hit
+      - stats-server.uri=http://stats-server:9090
 
   stats-db:
-    image: postgres:14-alpine
-    container_name: stat-db
+    image: postgres:14.6-alpine
+    container_name: stats_db
     ports:
       - "6541:5432"
     environment:
@@ -24,19 +25,19 @@ services:
   ewm-service:
     build: ./ewm-service
     image: ewm-service
-    container_name: ewm-service
+    container_name: ewm_service
     ports:
       - "8080:8080"
     depends_on:
-      - stats-server
       - ewm-db
+      - stats-server
     environment:
       - SPRING_DATASOURCE_URL=jdbc:postgresql://ewm-db:5432/ewm
-      - STATS_SERVICE_URL=http://stats-service:9090
+      - STATS-SERVER_URL=http://stat-server:9090
 
   ewm-db:
-    image: postgres:14-alpine
-    container_name: ewm-app-db
+    image: postgres:14.6-alpine
+    container_name: ewm_db
     ports:
       - "6542:5432"
     environment:
