+++ ewm-service/src/main/java/ru/practicum/utils/ConstantUtil.java
@@ -3,5 +3,6 @@ package ru.practicum.utils;
 public final class ConstantUtil {
 
     public static final String DATA_FORMAT = "yyyy-MM-dd HH:mm:ss";
+    public static final String EVENT_URI = "/events/";
 
 }
