+++ ewm-service/src/main/java/ru/practicum/utils/StatsUtil.java
@@ -1,0 +1,18 @@
+package ru.practicum.utils;
+
+import lombok.AllArgsConstructor;
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+import lombok.Setter;
+import ru.practicum.dto.StatsOutputDto;
+
+import java.util.List;
+
+@Getter
+@Setter
+@AllArgsConstructor
+@NoArgsConstructor
+public class StatsUtil {
+
+    private List<StatsOutputDto> stats;
+}
