+++ ewm-service/src/main/java/ru/practicum/utils/mapper/CompilationMapper.java
@@ -23,20 +23,22 @@ public final class CompilationMapper {
     }
 
     public static CompilationOutputDto toCompilationDto(Compilation compilation,
-                                                        Map<Event, List<Request>> confirmedRequests) {
+                                                        Map<Event, List<Request>> confirmedRequests,
+                                                        Map<Long, Long> views) {
         return CompilationOutputDto.builder()
                 .id(compilation.getId())
                 .title(compilation.getTitle())
                 .pinned(compilation.getPinned())
-                .events(EventMapper.toEventShortList(compilation.getEvents(), confirmedRequests))
+                .events(EventMapper.toEventShortList(compilation.getEvents(), confirmedRequests, views))
                 .build();
     }
 
     public static List<CompilationOutputDto> toCompDtoList(List<Compilation> comps,
-                                                           Map<Event, List<Request>> confirmedRequests) {
+                                                           Map<Event, List<Request>> confirmedRequests,
+                                                           Map<Long, Long> views) {
         List<CompilationOutputDto> compDto = new ArrayList<>();
         for (Compilation c : comps) {
-            compDto.add(toCompilationDto(c, confirmedRequests));
+            compDto.add(toCompilationDto(c, confirmedRequests, views));
         }
         return compDto;
     }
