--- stats-parent/stats-client/src/main/java/ru/practicum/client/StatClient.java
@@ -1,47 +1,0 @@
-package ru.practicum.client;
-
-import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.beans.factory.annotation.Value;
-import org.springframework.boot.web.client.RestTemplateBuilder;
-import org.springframework.http.ResponseEntity;
-import org.springframework.http.client.HttpComponentsClientHttpRequestFactory;
-import org.springframework.stereotype.Service;
-import org.springframework.web.util.DefaultUriBuilderFactory;
-import ru.practicum.dto.HitsInputDto;
-
-import java.time.LocalDateTime;
-import java.util.List;
-import java.util.Map;
-
-@Service
-public class StatClient extends BaseClient {
-
-    @Autowired
-    public StatClient(@Value("${stats-server.url}") String serverUrl, RestTemplateBuilder builder) {
-
-        super(builder
-                .uriTemplateHandler(new DefaultUriBuilderFactory(serverUrl))
-                .requestFactory(HttpComponentsClientHttpRequestFactory::new)
-                .build()
-        );
-    }
-
-    public ResponseEntity<Object> saveHit(HitsInputDto requestBody) {
-        return post("/hit", requestBody);
-    }
-
-    public ResponseEntity<Object> viewStats(LocalDateTime start,
-                                                  LocalDateTime end,
-                                                  List<String> uris,
-                                                  Boolean unique) {
-
-        Map<String, Object> parameters = Map.of(
-                "start", start,
-                "end", end,
-                "uris", uris,
-                "unique", unique
-        );
-        return get("/stats?start={start}&end={end}&uris={uris}&unique={unique}", parameters);
-    }
-
-}
