--- stats-parent/stats-service/Dockerfile
@@ -1,3 +1,0 @@
-FROM amazoncorretto:11-alpine-jdk
-COPY ./target/*.jar app.jar
-ENTRYPOINT ["java","-jar","/app.jar"]
