--- stats-parent/stats-service/src/main/resources/application.properties
@@ -1,12 +1,0 @@
-server.port=9090
-
-spring.jpa.hibernate.ddl-auto=none
-spring.jpa.properties.hibernate.dialect=org.hibernate.dialect.PostgreSQL10Dialect
-spring.jpa.properties.hibernate.format_sql=true
-spring.sql.init.mode=always
-
-#---
-spring.datasource.driverClassName=org.postgresql.Driver
-spring.datasource.url=jdbc:postgresql://localhost:5432/hit
-spring.datasource.username=root
-spring.datasource.password=root
