+++ stats/stats-client/src/main/java/ru/practicum/client/BaseClient.java
