+++ stats/stats-dto/src/main/java/ru/practicum/dto/HitsInputDto.java
