+++ stats/stats-dto/src/main/java/ru/practicum/dto/StatsOutputDto.java
