+++ stats/stats-service/Dockerfile
@@ -1,0 +1,4 @@
+FROM amazoncorretto:11-alpine-jdk
+ENV JAVA_TOOL_OPTIONS -agentlib:jdwp=transport=dt_socket,server=y,suspend=n,address=*:9091
+COPY target/*.jar app.jar
+ENTRYPOINT ["java","-jar","/app.jar"]
