+++ stats/stats-service/src/main/java/ru/practicum/stat/StatsServiceApp.java
@@ -1,0 +1,13 @@
+package ru.practicum.stat;
+
+import org.springframework.boot.SpringApplication;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+
+@SpringBootApplication
+public class StatsServiceApp {
+
+    public static void main(String[] args) {
+        SpringApplication.run(StatsServiceApp.class, args);
+    }
+
+}
