+++ stats/stats-service/src/main/java/ru/practicum/stat/controller/StatsController.java
