+++ stats/stats-service/src/main/java/ru/practicum/stat/model/Hit.java
@@ -1,0 +1,25 @@
+package ru.practicum.stat.model;
+
+import lombok.*;
+
+import javax.persistence.*;
+import java.time.LocalDateTime;
+
+@Entity
+@Table(name = "hits")
+@Getter
+@Setter
+@NoArgsConstructor
+@AllArgsConstructor
+@Builder
+public class Hit {
+
+    @Id
+    @GeneratedValue(strategy = GenerationType.IDENTITY)
+    private Long id;
+    private String app;
+    private String uri;
+    private String ip;
+    private LocalDateTime timestamp;
+
+}
