+++ stats/stats-service/src/main/java/ru/practicum/stat/repository/StatsRepository.java
@@ -1,0 +1,46 @@
+package ru.practicum.stat.repository;
+
+import org.springframework.data.jpa.repository.JpaRepository;
+import org.springframework.data.jpa.repository.Query;
+import org.springframework.stereotype.Repository;
+import ru.practicum.dto.StatsOutputDto;
+import ru.practicum.stat.model.Hit;
+
+import java.time.LocalDateTime;
+import java.util.List;
+
+@Repository
+public interface StatsRepository extends JpaRepository<Hit, Long> {
+
+    @Query("SELECT new ru.practicum.dto.StatsOutputDto(h.app, h.uri, COUNT(h.id)) " +
+            "FROM Hit h " +
+            "WHERE h.timestamp between ?1 AND ?2 " +
+            "GROUP BY h.app, h.uri " +
+            "ORDER BY COUNT (h.ip) DESC ")
+    List<StatsOutputDto> getAllStats(LocalDateTime start, LocalDateTime end);
+
+    @Query("SELECT new ru.practicum.dto.StatsOutputDto(h.app, h.uri, COUNT(DISTINCT h.ip)) " +
+            "FROM Hit h " +
+            "WHERE h.timestamp between ?1 AND ?2 " +
+            "GROUP BY h.app, h.uri " +
+            "ORDER BY COUNT (DISTINCT h.ip) DESC ")
+    List<StatsOutputDto> getAllStatsDistinctIp(LocalDateTime start, LocalDateTime end);
+
+    @Query("SELECT new ru.practicum.dto.StatsOutputDto(h.app, h.uri, COUNT(h.ip)) " +
+            "FROM Hit h " +
+            "WHERE h.timestamp between ?1 AND ?2 " +
+            "AND h.uri IN (?3) " +
+            "GROUP BY h.app, h.uri " +
+            "ORDER BY COUNT(h.ip) DESC ")
+    List<StatsOutputDto> getStats(LocalDateTime start, LocalDateTime end, List<String> uris);
+
+    @Query("SELECT new ru.practicum.dto.StatsOutputDto(h.app, h.uri, COUNT(DISTINCT h.ip)) " +
+            "FROM Hit h " +
+            "WHERE h.timestamp between ?1 AND ?2 " +
+            "AND h.uri IN (?3) " +
+            "GROUP BY h.app, h.uri " +
+            "ORDER BY COUNT(DISTINCT h.ip) DESC ")
+    List<StatsOutputDto> getStatsDistinctIp(LocalDateTime start, LocalDateTime end, List<String> uris);
+
+
+}
