+++ stats/stats-service/src/main/resources/application.properties
@@ -1,4 +1,4 @@
-server.port=8080
+server.port=9090
 
 stats-server.uri=http://localhost:9090
 
@@ -7,10 +7,8 @@ spring.jpa.properties.hibernate.dialect=org.hibernate.dialect.PostgreSQL10Dialec
 spring.jpa.properties.hibernate.format_sql=true
 spring.sql.init.mode=always
 
-logging.level.root=debug
-
 #---
 spring.datasource.driverClassName=org.postgresql.Driver
-spring.datasource.url=jdbc:postgresql://localhost:5432/ewm
+spring.datasource.url=jdbc:postgresql://localhost:5432/hit
 spring.datasource.username=root
 spring.datasource.password=root
